isolated casatools dependency to _casa modules.
