+++ docs/index.rst
@@ -11,7 +11,10 @@ Visread documentation
 .. |Tests badge| image:: https://github.com/MPoL-dev/visread/actions/workflows/tests.yml/badge.svg
    :target: https://github.com/MPoL-dev/visread/actions/workflows/tests.yml
 
-**What is this package**? This package is mostly a collection of documentation and tutorials demonstrating how to read visibilities (and associated metadata) from a calibrated CASA Measurement Set file into memory in a Python program using the built in CASA tools like `table` and `ms`. Once in memory, you can analyze them using your existing Python software stack, or save them to a binary file format (like ``*.npz`` or ``*.hdf5``) so that you can transport them to a computing environment free from a CASA dependency. 
+**What is this package**? This package is really three things:
+1) a collection of documentation and tutorials demonstrating how to read visibilities (and associated metadata) from a calibrated CASA Measurement Set file into memory using CASA tools like `table` and `ms`. For this, you don't even need to concern yourself with installing visread package, just browse the tutorials directly. 
+2) a few routines within the visread package that codify the data access patterns shown in the tutorials, available via `pip install visread[casa]` and described further in [installation.md]. These routines depend on the `casatools` package, and thus have the same Python version and operating system requirements that [Modular CASA](https://casadocs.readthedocs.io/en/stable/notebooks/introduction.html#Compatibility) does. Depending on the state of the release cycle, these can be quite constraining and preclude using modern Python versions.
+3) a set of non-imaging analysis and visualization routines to work with the visibility data, available via `pip install visread` and described further in [installation.md]. These core routines *do not* have a `casatools` dependency, and so they should be usuable in all current Python versions.
 
 **To get started**, you can get some ideas for working with `table` and `ms` directly from the *Introduction to CASA tools*. For convenience, you can install the *visread* package to provide some lightweight routines for common visibility manipulations, which are built on top of the core `casatools` functionality.
