+++ src/visread/visualization.py
@@ -1,31 +1,29 @@
 import numpy as np 
 import matplotlib.pyplot as plt
-import casatools 
 from . import utils, scatter, process
 
-ms = casatools.ms()
+def plot_baselines(u, v, title=None):
+    """
+    Make a plot of the baselines.
 
-def plot_baselines(filename, datadescid):
+    Args:
+        u: baseline in meters
+        v: baseline in meters.
+    
+    Returns:
+        matplotlib.fig
 
-    ms.open(filename)
-    ms.selectinit(datadescid=datadescid)
-    q = ms.getdata(["uvw"])
-    ms.selectinit(reset=True)
-    ms.close()
-
-    u, v, w = q["uvw"] * 1e-3  # [km]
-
+    """
     fig, ax = plt.subplots(nrows=1)
-    ax.scatter(u, v, s=0.5)
-    ax.set_title("DATA_DESC_ID: {:}".format(datadescid))
+    ax.scatter(u * 1e-3, v * 1e-3, s=0.5)
+    ax.set_title(title)
     ax.set_aspect("equal")
     ax.set_xlabel(r"$u$ [km]")
     ax.set_ylabel(r"$v$ [km]")
 
     return fig
 
-
-def plot_averaged_scatter(scatter, log=False, **kwargs):
+def plot_averaged_scatter(scatter, log=False,   **kwargs):
 
     xs = np.linspace(-5, 5)
 
@@ -45,60 +43,7 @@ def plot_averaged_scatter(scatter, log=False, **kwargs):
 
     return fig
 
-
-def plot_scatter_datadescid(
-    filename,
-    datadescid,
-    log=False,
-    sigma_rescale=1.0,
-    chan_slice=None,
-    apply_flags=True,
-    residual=True,
-    datacolumn="corrected_data"
-):
-    r"""
-    Plot a set of histograms of the scatter of the residual visibilities for the real and
-    imaginary components of the XX and YY polarizations.
-
-    Args:
-        filename (string): measurement set filename
-        datadescid (int): the DATA_DESC_ID to be queried
-        log (bool): should the histogram values be log scaled?
-        sigma_rescale (int):  multiply the uncertainties by this factor
-        chan_slice (slice): if not None, a slice object specifying the channels to subselect
-        apply_flags (bool): calculate the scatter *after* the flags have been applied
-        residual (bool): if True, subtract MODEL_DATA column (from a tclean model, most likely) to plot scatter of residual visibilities.
-        datacolumn (string): which datacolumn to use (i.e., 'corrected_data' or 'data').
-
-    Returns:
-        matplotlib figure with scatter histograms
-    """
-
-    if chan_slice is not None:
-        print("apply_flags setting is ignored when chan_slice is not None")
-
-        scatter_XX, scatter_YY = scatter.get_scatter_datadescid(
-            filename, datadescid, sigma_rescale, apply_flags=False, residual=residual, datacolumn=datacolumn
-        )
-
-        scatter_XX = scatter_XX[chan_slice]
-        scatter_YY = scatter_YY[chan_slice]
-
-    else:
-        scatter_XX, scatter_YY = scatter.get_scatter_datadescid(
-            filename, datadescid, sigma_rescale, apply_flags=apply_flags, residual=residual, datacolumn=datacolumn
-        )
-
-    scatter_XX = scatter_XX.flatten()
-    scatter_YY = scatter_YY.flatten()
-
-    fig = _scatter_hist(scatter_XX, scatter_YY, log=log)
-    fig.suptitle("DATA_DESC_ID: {:}".format(datadescid))
-
-    return fig
-
-
-def _scatter_hist(scatter_XX, scatter_YY, log=False, **kwargs):
+def scatter_hist(scatter_XX, scatter_YY, log=False, **kwargs):
     """
     Args:
         scatter_XX (1D numpy array)
@@ -139,18 +84,11 @@ def _scatter_hist(scatter_XX, scatter_YY, log=False, **kwargs):
     return fig
 
 
-def plot_weight_hist(filename, datadescid, log=False, **kwargs):
+def plot_weight_hist(weight_XX, weight_YY, title=None):
 
-    ms.open(filename)
-    ms.selectinit(datadescid=datadescid)
-    q = ms.getdata(["weight"])
-    ms.selectinit(reset=True)
-    ms.close()
-
-
-    weight_XX, weight_YY = q["weight"]
-    scatter_XX, scatter_YY = process.weight_to_sigma(q["weight"])
-
+    sigma_XX = process.weight_to_sigma(weight_XX)
+    sigma_YY = process.weight_to_sigma(weight_YY)
+    
     figsize = kwargs.get("figsize", (9, 9))
     bins = kwargs.get("bins", 20)
 
@@ -161,11 +99,11 @@ def plot_weight_hist(filename, datadescid, log=False, **kwargs):
     ax[0, 1].hist(weight_YY, bins=bins, log=log)
     ax[0, 1].set_xlabel(r"$w_\mathrm{YY}\,[1/\mathrm{Jy}^2]$")
 
-    ax[1, 0].hist(scatter_XX, bins=bins, log=log)
+    ax[1, 0].hist(sigma_XX, bins=bins, log=log)
     ax[1, 0].set_xlabel(r"$\sigma_\mathrm{XX}\,[\mathrm{Jy}]$")
-    ax[1, 1].hist(scatter_YY, bins=bins, log=log)
+    ax[1, 1].hist(sigma_YY, bins=bins, log=log)
     ax[1, 1].set_xlabel(r"$\sigma_\mathrm{YY}\,[\mathrm{Jy}]$")
     fig.subplots_adjust(hspace=0.25)
-    fig.suptitle("DATA_DESC_ID: {:}".format(datadescid))
+    fig.suptitle(title)
 
     return fig
